--- a/benchmarks_test.go
+++ b/benchmarks_test.go
@@ -3,7 +3,12 @@ package multiplex
 import (
 	"math/rand"
 	"net"
+	"sync"
+	"sync/atomic"
 	"testing"
+	"time"
+
+	"github.com/cevatbarisyilmaz/lossy"
 )
 
 func MakeLinearWriteDistribution() [][]byte {
@@ -34,6 +39,15 @@ func BenchmarkSmallPackets(b *testing.B) {
 	benchmarkPackets(b, msgs)
 }
 
+func BenchmarkSlowConnSmallPackets(b *testing.B) {
+	msgs := MakeSmallPacketDistribution()
+	pa, pb := net.Pipe()
+	la := lossy.NewConn(pa, 100*1024, 30*time.Millisecond, 100*time.Millisecond, 0.0, 48)
+	mpa := NewMultiplex(la, false)
+	mpb := NewMultiplex(pb, true)
+	benchmarkPacketsWithConn(b, msgs, mpa, mpb)
+}
+
 func BenchmarkLinearPackets(b *testing.B) {
 	msgs := MakeLinearWriteDistribution()
 	benchmarkPackets(b, msgs)
@@ -43,11 +57,15 @@ func benchmarkPackets(b *testing.B, msgs [][]byte) {
 	pa, pb := net.Pipe()
 	mpa := NewMultiplex(pa, false)
 	mpb := NewMultiplex(pb, true)
+	benchmarkPacketsWithConn(b, msgs, mpa, mpb)
+}
+
+func benchmarkPacketsWithConn(b *testing.B, msgs [][]byte, mpa, mpb *Multiplex) {
 	defer mpa.Close()
 	defer mpb.Close()
 
-	b.RunParallel(func(pb *testing.PB) {
-		receiveBuf := make([]byte, 2048)
+	streamPairs := make([][]*Stream, 0)
+	for i := 0; i < 64; i++ {
 		sa, err := mpa.NewStream()
 		if err != nil {
 			b.Error(err)
@@ -56,14 +74,46 @@ func benchmarkPackets(b *testing.B, msgs [][]byte) {
 		if err != nil {
 			b.Error(err)
 		}
+		streamPairs = append(streamPairs, []*Stream{sa, sb})
+	}
+
+	receivedBytes := uint64(0)
+	idx := int32(0)
+	b.ResetTimer()
+
+	b.RunParallel(func(pb *testing.PB) {
+		localIdx := atomic.AddInt32(&idx, 1)
+		if localIdx >= 64 {
+			panic("parallelism running into unallocated streams.")
+		}
+		localA := streamPairs[localIdx][0]
+		localB := streamPairs[localIdx][1]
+
+		var wg sync.WaitGroup
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			receiveBuf := make([]byte, 2048)
+
+			for {
+				n, err := localB.Read(receiveBuf)
+				if n == 0 || err != nil {
+					return
+				}
+				atomic.AddUint64(&receivedBytes, uint64(n))
+			}
+		}()
+
 		i := 0
 		for {
-			_, _ = sa.Write(msgs[i])
-			_, _ = sb.Read(receiveBuf)
+			_, _ = localA.Write(msgs[i])
 			i = (i + 1) % 1000
 			if !pb.Next() {
-				return
+				break
 			}
 		}
+		localA.Close()
+		wg.Wait()
 	})
+	b.SetBytes(int64(receivedBytes))
 }