--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,7 @@
 module github.com/libp2p/go-mplex
 
 require (
+	github.com/cevatbarisyilmaz/lossy v0.2.1
 	github.com/ipfs/go-log v1.0.2
 	github.com/libp2p/go-buffer-pool v0.0.2
 )