--- a/go.mod
+++ b/go.mod
@@ -4,6 +4,7 @@ require (
 	github.com/cevatbarisyilmaz/lossy v0.2.1
 	github.com/ipfs/go-log v1.0.2
 	github.com/libp2p/go-buffer-pool v0.0.2
+	github.com/libp2p/go-libp2p-core v0.5.1
 	google.golang.org/grpc v1.28.1
 )