--- a/go.mod
+++ b/go.mod
@@ -2,9 +2,13 @@ module github.com/libp2p/go-mplex
 
 require (
 	github.com/cevatbarisyilmaz/lossy v0.2.1
+	github.com/hashicorp/golang-lru v0.5.0 // indirect
 	github.com/ipfs/go-log v1.0.2
 	github.com/libp2p/go-buffer-pool v0.0.2
-	github.com/libp2p/go-libp2p-core v0.5.1
+	github.com/libp2p/go-libp2p-testing v0.1.2-0.20200422005655-8775583591d8
+	github.com/libp2p/go-msgio v0.0.4 // indirect
+	github.com/spacemonkeygo/openssl v0.0.0-20181017203307-c2dcc5cca94a // indirect
+	golang.org/x/crypto v0.0.0-20190618222545-ea8f1a30c443 // indirect
 	google.golang.org/grpc v1.28.1
 )