Work on #76 

Initially, running as `go test -bench . -benchtime 15s -cpu 1,2,4` yields
(Without write coalescing):
```
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-mplex
BenchmarkSmallPackets      	 7616295	      2520 ns/op
BenchmarkSmallPackets-2    	 3293174	      5364 ns/op
BenchmarkSmallPackets-4    	 2551482	      7120 ns/op
BenchmarkLinearPackets     	 5917447	      2975 ns/op
BenchmarkLinearPackets-2   	 2833734	      6439 ns/op
BenchmarkLinearPackets-4   	 2556615	      7125 ns/op
PASS
ok  	github.com/libp2p/go-mplex	145.408s
```

With Write Coalescing:
```
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-mplex
BenchmarkSmallPackets      	  171469	    104704 ns/op
BenchmarkSmallPackets-2    	  298460	     66740 ns/op
BenchmarkSmallPackets-4    	  490029	     37051 ns/op
BenchmarkLinearPackets     	  249435	     73914 ns/op
BenchmarkLinearPackets-2   	  463890	     45462 ns/op
BenchmarkLinearPackets-4   	  755593	     25591 ns/op
PASS
ok  	github.com/libp2p/go-mplex	122.683s
```

This is the expected failure mode though, since we're representing a 0-latency, infinite-bandwidth network.