@willscott 2020-04-16
updated benchmarks to consider a case where `Send()`'s on the underlying `net.Conn` block. (via a time.Sleep)

This causes somewhat unbelievable benchmarks at present:
```
w/o coalescing
BenchmarkSmallPackets               6505782	      2815 ns/op
BenchmarkSlowConnSmallPackets 	    4860	   3899876 ns/op
w/  coalescing
BenchmarkSmallPackets 	            14959983	  1494 ns/op
BenchmarkSlowConnSmallPackets 	    363711387	  41.3 ns/op
```