@Stebalien 2020-04-16
Hm, yeah, something is fishy.

Does the lossy packet conn you're using correctly simulate packet loss for a TCP connection? That is, we don't expect data to actually be _lost_, we just expect things to slow down due to retransmissions.