@willscott 2020-04-16
The wrapper around the net.Conn blocks the sending thread with a Sleep based on packet size (to cap bandwidth) and latency. It is configured in these benchmarks to not drop any packets.