@willscott 2020-04-17
ok, some more reasonable numbers: (a 100kbps link with 30ms latency)
```
with write coalescing

BenchmarkSlowConnSmallPackets     	    1204	  23187782 ns/op	  15.87 MB/s
BenchmarkSlowConnSmallPackets-2   	    1002	  25399463 ns/op	  13.15 MB/s
BenchmarkSlowConnSmallPackets-4   	     834	  45408887 ns/op	   5.43 MB/s

without

BenchmarkSlowConnSmallPackets     	     982	  23468765 ns/op	  12.66 MB/s
BenchmarkSlowConnSmallPackets-2   	    1172	  23898852 ns/op	  15.07 MB/s
BenchmarkSlowConnSmallPackets-4   	     782	  26059704 ns/op	  10.15 MB/s
```

Interesting that the 4-core (more contention) test starts falling off link saturation in the write coalescing case. I'll re-run with a faster network and see how that trends.