@willscott 2020-04-27
This is now using https://github.com/libp2p/go-libp2p-testing/pull/21

I believe the top numbers are actually due to imprecision in how go bench reports work done by parallel threads vs overall. I now no longer believe any meaningful change is incurred in mplex from write coalescing.