@willscott 2020-04-29
for comparison, the current local output
(with coalescing)
```
CPU Bound Limit:   395835	      3079 ns/op	39079050.22 MB/s
Latency Bound Limit: 16ms
Bandwidth Bound Limit: 102400Kbps
Network Bound Limit:   163454	      7607 ns/op	6537690.62 MB/s
At MaxProc 1 102400Kbps / 8ms latency:    53253	     23214 ns/op	697423.50 MB/s
At MaxProc 2 102400Kbps / 8ms latency:    53114	     23204 ns/op	695911.14 MB/s
At MaxProc 4 102400Kbps / 8ms latency:    52712	     23223 ns/op	690329.83 MB/s
Slowdown is 0.723299%
PASS
```
(without)
```
CPU Bound Limit:    60412	     19440 ns/op	945400.60 MB/s
Latency Bound Limit: 64ms
Bandwidth Bound Limit: 102400Kbps
Network Bound Limit:    40866	     28791 ns/op	432383.89 MB/s
At MaxProc 1 102400Kbps / 32ms latency:    52602	     25348 ns/op	630851.91 MB/s
At MaxProc 2 102400Kbps / 32ms latency:    56181	     23034 ns/op	741973.43 MB/s
At MaxProc 4 102400Kbps / 32ms latency:    55123	     23116 ns/op	725004.12 MB/s
Slowdown is 1.940173%
PASS
```