@Stebalien 2020-04-30
Feedback from sync call:

* It would be nice to get numbers on how many writes we're doing (an approximation of packets sent).
* It would be nice to re-run these tests with jitter.

Write coalescing is targeting jitter, packet-loss, and syscall overhead.