@willscott 2020-05-04
> @willscott could you re-disable write-coalescing so we can get these merged?

rebased to remove